% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dten.R
\name{getAllNetStats}
\alias{getAllNetStats}
\title{\code{getAllNetStats} Gets all the summary stats about the network
without restricting to unique nodes (like \code{getNetSummaries} does)
Also, we no longer care about enrichment terms}
\usage{
getAllNetStats(netlist, synids = NULL)
}
\arguments{
\item{list}{of objects from network calls}
}
\value{
list of node tables
}
\description{
\code{getAllNetStats} Gets all the summary stats about the network
without restricting to unique nodes (like \code{getNetSummaries} does)
Also, we no longer care about enrichment terms
}
