% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcsf.R
\name{getDrugsFromGraph}
\alias{getDrugsFromGraph}
\title{\code{getDrugsFromGraph} Identifies drugs in a}
\usage{
getDrugsFromGraph(drug.graph)
}
\arguments{
\item{combined.graph}{Graph}
}
\value{
drug names
}
\description{
\code{getDrugsFromGraph} Identifies drugs in a
}
