% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcsf.R
\name{getDrugTargets}
\alias{getDrugTargets}
\title{\code{getDrugTargets} Gets drug target graph and returns table of drug targeets}
\usage{
getDrugTargets(drug.graph)
}
\value{
tidied dataset
}
\description{
\code{getDrugTargets} Gets drug target graph and returns table of drug targeets
}
