% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaviper.R
\name{getViperForCondition}
\alias{getViperForCondition}
\title{\code{getViperForCondition} takes a matrix from viper and condition of interest aind computes differential reg}
\usage{
getViperForCondition(v.res, condition, pvalthresh = 0.05,
  useEntrez = TRUE, p.corr = TRUE)
}
\arguments{
\item{condition}{set of columns}
}
\value{
list
}
\description{
\code{getViperForCondition} takes a matrix from viper and condition of interest aind computes differential reg
}
