% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtex.R
\name{renameDrugIds}
\alias{renameDrugIds}
\title{\code{renameDrugIds} Remaps drug ids to drug names for view-ability}
\usage{
renameDrugIds(pcsf.res, dummies)
}
\description{
\code{renameDrugIds} Remaps drug ids to drug names for view-ability
}
