% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcsf.R
\name{runPcsfWithParams}
\alias{runPcsfWithParams}
\title{\code{runPcsfWithParams} Identifies drugs in a}
\usage{
runPcsfWithParams(ppi, terminals, dummies, w = 2, b = 1, mu = 5e-04,
  doRand = FALSE)
}
\description{
\code{runPcsfWithParams} Identifies drugs in a
}
\keyword{PCSF}
\keyword{network}
\keyword{terminals}
