% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcsf.R
\name{shuffleTerminalsAndRun}
\alias{shuffleTerminalsAndRun}
\title{\code{shuffleTerminalsAndRun} will shuffle terminals for a permutation test}
\usage{
shuffleTerminalsAndRun(ppi, terminals, dummies, w, b, mu, numShuffles)
}
\value{
original network as well as p-values on each node
}
\description{
\code{shuffleTerminalsAndRun} will shuffle terminals for a permutation test
}
